Pep8 cleanup
