+++ ooiui/core/routes/science.py
@@ -6,37 +6,22 @@ Defines the application routes
 '''
 from ooiui.core.app import app
 from flask import request, render_template, Response, jsonify
-from flask import stream_with_context, make_response, redirect, url_for
-from ooiui.core.routes.common import get_login
-
-
+from flask import stream_with_context
+from ooiui.core.routes.common import get_login, login_required
 import requests
-import os
-import json
-from datetime import datetime,timedelta
-import time
-import numpy as np
-import math
 import urllib2
-from functools import wraps
 
-
-def login_required(f):
-    @wraps(f)
-    def decorated_function(*args, **kwargs):
-        if get_login() is None:
-            return redirect(url_for('new_index'))
-        return f(*args, **kwargs)
-    return decorated_function
 
 @app.route('/')
 def new_index():
     urllib2.urlopen(app.config['GOOGLE_ANALYTICS_URL'] + '&dp=%2Fscience%2Fworld')
     return render_template('science/index.html')
 
+
 @app.route('/landing/pioneer')
 def landing_pioneer():
     return render_template('landing/pioneer.html')
+
 
 @app.route('/assets/list')
 @app.route('/assets/list/')
@@ -44,25 +29,28 @@ def instr_index():
     urllib2.urlopen(app.config['GOOGLE_ANALYTICS_URL'] + '&dp=%2Fassets')
     return render_template('asset_management/assetslist.html')
 
+
 @app.route('/events/list/')
 def event_list():
     urllib2.urlopen(app.config['GOOGLE_ANALYTICS_URL'] + '&dp=%2Fevents')
     return render_template('asset_management/eventslist.html')
 
+
 @app.route('/event/<int:id>', methods=['GET'])
 def event_index(id):
-    #?id=%s' % id
-    return render_template('asset_management/event.html',id=id)
+    return render_template('asset_management/event.html', id=id)
+
 
 @app.route('/event/<string:new>/<int:aid>/<string:aclass>', methods=['GET'])
-def event_new(new,aid,aclass):
-    #?id=%s' % id
-    return render_template('asset_management/event.html',id=str(new),assetid=aid,aclass=str(aclass))
+def event_new(new, aid, aclass):
+    return render_template('asset_management/event.html', id=str(new), assetid=aid, aclass=str(aclass))
+
 
 @app.route('/streams/')
 def streams_page():
     urllib2.urlopen(app.config['GOOGLE_ANALYTICS_URL'] + '&dp=%2Fstreams')
     return render_template('science/streams.html')
+
 
 @app.route('/plotting', methods=['GET'])
 @app.route('/plotting/', methods=['GET'])
@@ -71,11 +59,13 @@ def show_plotting_no_path():
     urllib2.urlopen(app.config['GOOGLE_ANALYTICS_URL'] + '&dp=%2Fplotting')
     return plotting_page(None)
 
+
 @app.route('/plotting/<path:path>', methods=['GET'])
 def plotting_page(path):
     urllib2.urlopen(app.config['GOOGLE_ANALYTICS_URL'] + '&dp=%2Fplotting')
     print path
     return render_template('science/plotting.html')
+
 
 @app.route('/getdata/')
 def getData():
@@ -84,11 +74,6 @@ def getData():
     '''
     instr = request.args['instrument']
     stream = request.args['stream']
-
-    #std = request.args['startdate']
-    #edd = request.args['enddate']
-    #param = request.args['variables']
-    #ann = request.args['annotaton']
     ann = "?annotation=true"
 
     response = requests.get(app.config['SERVICES_URL'] + '/uframe/get_data'+"/"+instr+"/"+stream+ann, params=request.args)
@@ -104,12 +89,12 @@ def getUframeDataProxy():
     try:
         instr = request.args['instrument']
         stream = request.args['stream']
-        #comma list
+        # comma list
         xvars = request.args['xvars']
         yvars = request.args['yvars']
 
-        #there should be a start and end date in the params
-        #?startdate=2015-01-21T22:01:48.103Z&enddate=2015-04-29T10:10:51.563Z
+        # there should be a start and end date in the params
+        # ?startdate=2015-01-21T22:01:48.103Z&enddate=2015-04-29T10:10:51.563Z
 
         data_url = "/".join([app.config['SERVICES_URL'],'uframe/get_data',instr,stream,xvars,yvars])        
         response = requests.get(data_url, params=request.args)
@@ -118,6 +103,7 @@ def getUframeDataProxy():
         return data_text, response.status_code, dict(response.headers)
     except Exception,e:
         return jsonify(error=str(e))
+
 
 @app.route('/api/annotation', methods=['GET'])
 def get_annotations():
@@ -145,10 +131,12 @@ def put_annotation(id):
     return response.text, response.status_code
 
 #old
+
 @app.route('/api/array')
 def array_proxy():
     response = requests.get(app.config['SERVICES_URL'] + '/arrays', params=request.args)
     return response.text, response.status_code
+
 
 @app.route('/api/uframe/get_structured_toc')
 def structured_toc_proxy():
@@ -156,10 +144,12 @@ def structured_toc_proxy():
     return response.text, response.status_code
 
 #old
+
 @app.route('/api/platform_deployment')
 def platform_deployment_proxy():
     response = requests.get(app.config['SERVICES_URL'] + '/platform_deployments', params=request.args)
     return response.text, response.status_code
+
 
 @app.route('/api/display_name')
 def display_name():
@@ -168,20 +158,24 @@ def display_name():
     return response.text, response.status_code
 
 #Assets
+
 @app.route('/api/asset_deployment', methods=['GET'])
 def instrument_deployment_proxy():
     response = requests.get(app.config['SERVICES_URL'] + '/uframe/assets', params=request.args)
     return response.text, response.status_code
+
 
 @app.route('/api/asset_deployment/<int:id>', methods=['GET'])
 def instrument_deployment_get(id):
     response = requests.get(app.config['SERVICES_URL'] + '/uframe/assets/%s' % id, data=request.data)
     return response.text, response.status_code
 
+
 @app.route('/api/asset_deployment/<int:id>', methods=['PUT'])
 def instrument_deployment_put(id):
     response = requests.put(app.config['SERVICES_URL'] + '/uframe/assets/%s' % id, data=request.data)
     return response.text, response.status_code
+
 
 @app.route('/api/asset_deployment', methods=['POST'])
 def instrument_deployment_post():
@@ -189,31 +183,37 @@ def instrument_deployment_post():
     return response.text, response.status_code
 
 #not working/using now
+
 @app.route('/api/asset_deployment/<int:id>', methods=['DELETE'])
 def instrument_deployment_delete(id):
     response = requests.delete(app.config['SERVICES_URL'] + '/uframe/assets/%s' % id, data=request.data)
     return response.text, response.status_code
 
 #Events
+
 @app.route('/api/asset_events', methods=['GET'])
 def event_deployments_proxy():
     response = requests.get(app.config['SERVICES_URL'] + '/uframe/events', params=request.args)
     return response.text, response.status_code
+
 
 @app.route('/api/asset_events/<int:id>', methods=['GET'])
 def event_deployment_get(id):
     response = requests.get(app.config['SERVICES_URL'] + '/uframe/events/%s' % id, params=request.args)
     return response.text, response.status_code
 
+
 @app.route('/api/asset_events/<int:id>', methods=['PUT'])
 def asset_event_put(id):
     response = requests.put(app.config['SERVICES_URL'] + '/uframe/events/%s' % id, data=request.data)
     return response.text, response.status_code
 
+
 @app.route('/api/asset_events', methods=['POST'])
 def asset_event_post():
     response = requests.post(app.config['SERVICES_URL'] + '/uframe/events', data=request.data)
     return response.text, response.status_code
+
 
 @app.route('/api/events', methods=['GET'])
 def get_event_by_ref_des():
@@ -226,11 +226,13 @@ def op_log():
     urllib2.urlopen(app.config['GOOGLE_ANALYTICS_URL'] + '&dp=%2FopLog')
     return render_template("common/opLog.html")
 
+
 @app.route('/api/uframe/stream')
 def stream_proxy():
     token = get_login()
     response = requests.get(app.config['SERVICES_URL'] + '/uframe/stream', auth=(token, ''), params=request.args)
     return response.text, response.status_code
+
 
 @app.route('/api/uframe/get_metadata/<string:stream_name>/<string:reference_designator>', methods=['GET'])
 def metadata_proxy(stream_name,reference_designator):
@@ -241,6 +243,7 @@ def metadata_proxy(stream_name,reference_designator):
     response = requests.get(app.config['SERVICES_URL'] + '/uframe/get_metadata/%s/%s' % (stream_name, reference_designator), auth=(token, ''), params=request.args)
     return response.text, response.status_code
 
+
 @app.route('/api/uframe/get_metadata_times/<string:stream_name>/<string:reference_designator>', methods=['GET'])
 def metadata_times_proxy(stream_name,reference_designator):
     '''
@@ -250,6 +253,7 @@ def metadata_times_proxy(stream_name,reference_designator):
     response = requests.get(app.config['SERVICES_URL'] + '/uframe/get_metadata_times/%s/%s' % (stream_name, reference_designator), auth=(token, ''), params=request.args)
     return response.text, response.status_code
 
+
 @app.route('/api/uframe/get_csv/<string:stream_name>/<string:reference_designator>/<string:start>/<string:end>')
 def get_csv(stream_name, reference_designator,start,end):
     token = get_login()
@@ -257,6 +261,7 @@ def get_csv(stream_name, reference_designator,start,end):
     url = app.config['SERVICES_URL'] + '/uframe/get_csv/%s/%s/%s/%s/%s' % (stream_name, reference_designator,start,end,dpa)
     req = requests.get(url, auth=(token, ''), stream=True)
     return Response(stream_with_context(req.iter_content(chunk_size=1024*1024*4)), headers=dict(req.headers))
+
 
 @app.route('/api/uframe/get_json/<string:stream_name>/<string:reference_designator>/<string:start>/<string:end>/<string:provenance>/<string:annotations>')
 def get_json(stream_name, reference_designator, start, end, provenance, annotations):
@@ -266,6 +271,7 @@ def get_json(stream_name, reference_designator, start, end, provenance, annotati
     req = requests.get(url, auth=(token, ''), stream=True,params=request.args)
     return Response(stream_with_context(req.iter_content(chunk_size=1024*1024*4)), headers=dict(req.headers))
 
+
 @app.route('/api/uframe/get_netcdf/<string:stream_name>/<string:reference_designator>/<string:start>/<string:end>/<string:provenance>/<string:annotations>')
 def get_netcdf(stream_name, reference_designator, start, end, provenance, annotations):
     token = get_login()
@@ -274,11 +280,13 @@ def get_netcdf(stream_name, reference_designator, start, end, provenance, annota
                        % (stream_name, reference_designator, start, end, dpa, provenance, annotations), auth=(token, ''), stream=True)
     return Response(stream_with_context(req.iter_content(chunk_size=1024*1024*4)), headers=dict(req.headers))
 
+
 @app.route('/api/uframe/get_profiles/<string:stream_name>/<string:reference_designator>')
 def get_profiles(stream_name, reference_designator):
     token = get_login()
     req = requests.get(app.config['SERVICES_URL'] + '/uframe/get_profiles/%s/%s/%s/%s' % (stream_name, reference_designator), auth=(token, ''), stream=True)
     return Response(stream_with_context(req.iter_content(chunk_size=1024*1024*4)), headers=dict(req.headers))
+
 
 @app.route('/svg/plot/<string:instrument>/<string:stream>', methods=['GET'])
 def get_plotdemo(instrument, stream):
@@ -288,8 +296,9 @@ def get_plotdemo(instrument, stream):
     req = requests.get(app.config['SERVICES_URL'] + '/uframe/plot/%s/%s' % (instrument, stream), auth=(token, ''), params=request.args)
     t1 = time.time()
     print "GUI took %s" % (t1 - t0)
-     # they fake the response to 200
+    # they fake the response to 200
     return req.content, req.status_code, dict(req.headers)
+
 
 # C2 Routes
 @app.route('/api/c2/array_display/<string:array_code>', methods=['GET'])
@@ -297,15 +306,18 @@ def get_c2_array_display(array_code):
     response = requests.get(app.config['SERVICES_URL'] + '/c2/array_display/%s' % (array_code))
     return response.text, response.status_code
 
+
 @app.route('/api/c2/platform_display/<string:reference_designator>', methods=['GET'])
 def get_c2_platform_display(reference_designator):
     response = requests.get(app.config['SERVICES_URL'] + '/c2/platform_display/%s' % (reference_designator))
     return response.text, response.status_code
 
+
 @app.route('/api/c2/instrument_display/<string:reference_designator>', methods=['GET'])
 def get_c2_instrument_display(reference_designator):
     response = requests.get(app.config['SERVICES_URL'] + '/c2/instrument_display/%s' % (reference_designator))
     return response.text, response.status_code
+
 
 @app.route('/api/c2/instrument/<string:reference_designator>/<string:stream_name>', methods=['GET'])
 def get_c2_instrument_fields(reference_designator, stream_name):
